// src/api/openapi/models/UpdateCommunityFarmRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateCommunityFarmRequest = {
    farm_name?: string;
    location?: string;
    description?: string;
    district?: string;
    size?: string;
    avatar?: Blob;
    cover_photo?: Blob;
};